Pull back comments api
